web/requirements.txt
Flask==2.3.3
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
numpy>=1.21.0
Werkzeug>=2.3.0
